Memory optimization
Added a RuleStorage abstraction that is used to store rules in a serialized form.
The idea is to be able to operate with rules identifiers and keep them in a smaller
serialized form. When we need a rule, it is deserialized, and saved to an in-memory map.

Additionally, we are able to store serialized rules in a file, and read them on-demand.
This allows us to save quite a lot of memory, but in exchange, filters parsing becomes very slow.